requirements.txt
anyio==3.6.1
astroid==2.12.5
attrs==22.1.0
autopep8==1.7.0
base58==2.1.1
black==22.8.0
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
commonmark==0.9.1
dill==0.3.5.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pydocstyle==6.1.1
Pygments==2.13.0
pylint==2.15.0
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.1.3
PyYAML==6.0
rfc3986==1.5.0
rich==12.5.1
sniffio==1.3.0
snowballstemmer==2.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
wrapt==1.14.1
